// cli/export-data/src/tasks/sams/sams-raw.types.ts
/* * */

export interface SamExport {
	_id: number
	agency_id: string
	apex_version: null | string
	device_id: null | string
	end_time: null | string
	first_transaction_ase_counter_value: null | number
	first_transaction_id: null | string
	first_transaction_type: null | string
	last_transaction_ase_counter_value: null | number
	last_transaction_id: null | string
	last_transaction_type: null | string
	start_time: null | string
	transactions_expected: null | number
	transactions_found: null | number
	transactions_missing: null | number
	vehicle_id: null | number
}